Move tag command to App.cs, leave window open
